Add the `embedded_brightness_search` parameter to the camera driver
The `embedded_brightness_search` parameter is used to switch between the
camera-embedded brightness search and the driver-implemented brightness
search. The camera-embedded brightness search doesn't work for all
encodings because the computation of the brightness is different between
Pylon library and the pylon-ros-camera driver.
The embedded brightness search could be used always by using the
`AutoBrightnessStatus` data chunk, but the implementation is more complex.

In the process, change the `binary_exp_search_` member to an
`std::unique_ptr` rather than a raw pointer.

Signed-off-by: Gaël Écorchard <gael@km-robotics.cz>